+++ benchmarks/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @tanstack/virtual-benchmarks
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.14.4
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`2ba5eb6`](https://github.com/TanStack/virtual/commit/2ba5eb60f108f4ba9b2bd9570bbd41f9ce618438)]:
+  - @tanstack/react-virtual@3.14.3
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.14.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.14.1
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`73e115d`](https://github.com/TanStack/virtual/commit/73e115d53faf01f073bc4a7bd71ee0139307f4a8)]:
+  - @tanstack/react-virtual@3.14.0
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.13.26
+
 ## 0.0.1
 
 ### Patch Changes